apisGabiMakeup/requirements.txt
asgiref==3.8.1
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.15.1
drf-yasg==1.21.7
google-cloud==0.34.0
gunicorn==21.2.0
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
